+++ rubato/struct/gameobject/sprites/animation.py
@@ -21,7 +21,7 @@ class Animation(Component):
     Args:
         scale: The scale of the animation. Defaults to Vector(1, 1).
         fps: The frames per second of the animation. Defaults to 24.
-        anti_aliasing: Whether to use anti-aliasing on the animation. Defaults to False.
+        aa: Whether to use anti-aliasing on the animation. Defaults to False.
         flipx: Whether to flip the animation horizontally. Defaults to False.
         flipy: Whether to flip the animation vertically. Defaults to False.
         offset: The offset of the animation from the game object. Defaults to Vector(0, 0).
@@ -33,7 +33,7 @@ class Animation(Component):
         self,
         scale: Vector = Vector(1, 1),
         fps: int = 24,
-        anti_aliasing: bool = False,
+        aa: bool = False,
         flipx: bool = False,
         flipy: bool = False,
         offset: Vector = Vector(),
@@ -60,7 +60,7 @@ class Animation(Component):
         """Whether the animation should loop."""
         self.scale: Vector = scale
         """The scale of the animation."""
-        self.aa: bool = anti_aliasing
+        self.aa: bool = aa
         """Whether or not to enable anti aliasing."""
         self.flipx: bool = flipx
         """Whether or not to flip the animation along the x axis."""
@@ -91,7 +91,7 @@ class Animation(Component):
         self.animation_frames_left = len(self._states[self.current_state]) - (1 + self._current_frame)
 
     @property
-    def image(self) -> sdl2.surface.SDL_Surface:
+    def image(self) -> sdl2.SDL_Surface:
         """The current SDL Surface holding the image."""
         return self._states[self.current_state][self.current_frame].image
 
@@ -110,7 +110,6 @@ class Animation(Component):
 
         sprite.scale = calculated_scale
         # pylint: disable=protected-access
-        sprite._update_rotozoom()
         return sprite
 
     def set_current_state(self, new_state: str, loop: bool = False, freeze: int = -1):
@@ -237,12 +236,6 @@ class Animation(Component):
 
         self.add(state_name, state)
 
-    def update(self):
-        if self.hidden:
-            return
-
-        self.anim_frame.update()
-
     def draw(self, camera: Camera):
         """Draws the animation frame and steps the animation forward."""
         if self.hidden:
@@ -254,7 +247,7 @@ class Animation(Component):
             self.anim_tick()
             self._time_count -= self._time_step
 
-        Draw.queue_sprite(
+        Draw.queue_surf(
             self.anim_frame, camera.transform((self.gameobj.pos + self.offset) - self.anim_frame.get_size() / 2),
             self.true_z
         )
@@ -282,7 +275,7 @@ class Animation(Component):
         new = Animation(
             scale=self.scale,
             fps=self.fps,
-            anti_aliasing=self.aa,
+            aa=self.aa,
             flipx=self.flipx,
             flipy=self.flipy,
             offset=self.offset,